Enum amethyst_engine::Trans

pub enum Trans {
    None,
    Pop,
    Push(Box<State>),
    Switch(Box<State>),
    Quit,
}

Types of state transitions.

Variants

None

Continue as normal.

Pop

Remove the active state and resume the next state on the stack or stop if there are none.

Push

Pause the active state and push a new state onto the stack.

Switch

Remove the current state on the stack and insert a different one.

Quit

Stop and remove all states and shut down the engine.